+++ src/buffer/message_view.rs
@@ -139,13 +139,11 @@ impl<'a> ChannelQueryLayout<'a> {
             .into_iter()
             .flatten()
             .find(|current_user| *current_user == user);
-        let is_user_offline = if self.target.is_query() {
-            false
-        } else {
-            match self.config.buffer.nickname.shown_status {
-                ShownStatus::Current => user_in_channel.is_none(),
-                ShownStatus::Historical => false,
+        let is_user_offline = match self.config.buffer.nickname.shown_status {
+            ShownStatus::Current => {
+                self.target.is_channel() && user_in_channel.is_none()
             }
+            ShownStatus::Historical => false,
         };
 
         let nickname_style = theme::selectable_text::nickname(